move ticketing types into ticketingSlice
