+++ src/features/ticketing/ticketingSlice.ts
@@ -1,8 +1,43 @@
 import { createSlice, createAsyncThunk, PayloadAction, CaseReducer } from '@reduxjs/toolkit'
 import { RootState } from '../../app/store'
-import { CartItem, Play, Ticket, ticketingState } from './ticketingTypes'
 import format from "date-fns/format";
 import { titleCase } from '../../utils'
+
+export interface CartItem {
+    product_id: number,     // references state.tickets.eventid
+    qty: number,
+    name: string,
+    desc: string,
+    product_img_url: string,
+    price: number,
+}
+
+export interface Ticket {
+    eventid: number,             // references showtime.id in database
+    playid: string,
+    admission_type: 'General Admission',
+    date: Date,
+    ticket_price: number,
+    concession_price: number,
+    availableseats: number,
+}
+
+export interface Play {
+    id: string,
+    title: string,
+    description: string,
+    image_url: string,
+}
+
+export type LoadStatus = 'idle' | 'loading' | 'success' | 'failed'
+export interface ticketingState {
+    cart: CartItem[],
+    tickets: Ticket[],
+    plays: Play[],
+    status: LoadStatus,
+}
+
+
 
 const fetchData = async (url: string) => {
     try {
